split up directory resolver from indexing
